fst2: Add 0.1ms delay after reading touch FIFO event
This delay will prevent the firmware from not being able to reset/update the FIFO in time for the next event.
In the worst case, the total delay time (10ms) will exceed 8.3ms
(= 1/120Hz), so I change the value of MAX_FIFO_EVENT from 100 to 32, it is confirmed by vendor.

Bug: 254923538
Test: Build pass
Signed-off-by: Mason Wang <masonwang@google.com>
Change-Id: I3b4219ec593720b709f6082b2f60a7b9c35a6b9d